# api/bundled/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "description": "Egolia API",
    "version": "0.0.0",
    "title": "Egolia API",
    "license": {
      "name": "MIT",
      "url": "https://github.com/egolia-uit/egolia/blob/main/LICENSE"
    },
    "contact": {
      "name": "Egolia Support",
      "email": "egolia@googlegroups.com"
    },
    "x-apiName": "api"
  },
  "servers": [
    {
      "url": "http://api.egolia.localhost"
    }
  ],
...
  "paths": {
...
    "/billing/courses/{courseId}/checkout": {
      "post": {
        "security": [
          {
            "billing_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "billing_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Billing"
        ],
        "summary": "Checkout course",
        "operationId": "checkoutCourse",
        "description": "- authenticated:\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/billing_courseIdPath"
          }
        ],
        "responses": {
          "201": {
            "description": "Checkout successful, transaction created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactionId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "0f4f19b1-1602-4c12-ac1a-9ad5baf1e4dc"
                    },
                    "paymentUrl": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://api.egolia.com/billing/transactions/0f4f19b1-1602-4c12-ac1a-9ad5baf1e4dc"
                    }
                  },
                  "required": [
                    "transactionId",
                    "paymentUrl"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/billing_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/billing_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/billing_ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/billing_InternalServerError"
          }
        }
      }
    },
    "/billing/admin/analytics/revenue": {
      "get": {
        "security": [
          {
            "billing_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "billing_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Billing"
        ],
        "summary": "Get platform revenue analytics",
        "operationId": "getPlatformRevenueAnalytics",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-03-01"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-03-31"
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue analytics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/billing_RevenueAnalytics"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/billing_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/billing_ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/billing_InternalServerError"
          }
        }
      }
    },
    "/billing/transactions": {
      "get": {
        "security": [
          {
            "billing_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "billing_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Billing"
        ],
        "summary": "Get transactions",
        "operationId": "getTransactions",
        "parameters": [
          {
            "$ref": "#/components/parameters/billing_pageQuery"
          },
          {
            "$ref": "#/components/parameters/billing_limitQuery"
          },
          {
            "$ref": "#/components/parameters/billing_orderQuery"
          },
          {
            "name": "courseId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "learnerId",
            "in": "query",
            "description": "If requested by non admin, then it will be that user",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/billing_id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Learner billing history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/billing_Transaction"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/billing_Pagination"
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/billing_UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/billing_InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "course_OIDC": {
        "type": "openIdConnect",
        "openIdConnectUrl": "http://authentik.egolia.localhost/application/o/api-web/.well-known/openid-configuration",
        "description": "OpenID Connect Security Scheme",
        "x-defaultClientId": "api-web",
        "x-scalar-client-id": "api-web"
      },
      "course_Oauth2": {
        "type": "oauth2",
        "description": "OAuth2 Security Scheme",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://authentik.notopia.localhost/application/o/authorize/",
            "tokenUrl": "http://authentik.notopia.localhost/application/o/token/",
            "scopes": {
              "openid": "OpenID Connect scope",
              "email": "Email scope",
              "profile": "Profile scope",
              "entitlements": "Entitlements scope"
            },
            "x-defaultClientId": "api-web",
            "x-scalar-client-id": "api-web",
            "x-usePkce": "SHA-256",
            "x-scalar-credentials-location": "header"
          },
          "password": {
            "tokenUrl": "http://authentik.notopia.localhost/application/o/token/",
            "scopes": {
              "openid": "OpenID Connect scope",
              "email": "Email scope",
              "profile": "Profile scope",
              "entitlements": "Entitlements scope"
            },
            "x-default-scopes": [
              "openid",
              "profile",
              "email",
              "entitlements"
            ],
            "x-defaultClientId": "api-web",
...
    "parameters": {
...
      "billing_courseIdPath": {
        "name": "courseId",
        "in": "path",
        "required": true,
        "description": "Unique identifier of the course",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "example": "834897ce-45e7-43a5-92b2-4ec1f0dc3cd1"
      },
      "billing_pageQuery": {
        "name": "page",
        "in": "query",
        "description": "Page number for pagination",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "example": 1
      },
      "billing_limitQuery": {
        "name": "limit",
        "in": "query",
        "description": "Number of items per page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "example": 20
      },
      "billing_orderQuery": {
        "name": "order",
        "in": "query",
        "description": "Sort order",
        "schema": {
          "type": "string",
          "enum": [
            "asc",
            "desc"
          ],
          "default": "desc",
          "example": "desc"
        }
      },
...
    },
    "schemas": {
...
      "billing_Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "EXAMPLE_CODE"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Example error"
          },
          "more_info": {
            "type": "string",
            "description": "URL with more information about the error",
            "example": "https://api.egolia.com/docs/errors/EXAMPLE_CODE"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "billing_RevenueAnalytics": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-22T10:30:00Z"
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-22T10:30:00Z"
          },
          "totalRevenue": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "example": 152000000
          },
          "completedTransactions": {
            "type": "integer",
            "minimum": 0,
            "example": 421
          },
          "failedTransactions": {
            "type": "integer",
            "minimum": 0,
            "example": 9
          }
        },
        "required": [
          "from",
          "to",
          "totalRevenue",
          "completedTransactions",
          "failedTransactions"
        ]
      },
      "billing_id": {
        "type": "string",
        "format": "number",
        "description": "User ID from Authentik (need to change subject mode to User's ID instead of hashed)",
        "example": "100"
      },
      "billing_TransactionStatus": {
        "type": "string",
        "description": "Current status of a billing transaction",
        "enum": [
          "pending",
          "completed",
          "failed"
        ],
        "example": "completed"
      },
      "billing_Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "4c7c8f0a-4f6e-4da1-867d-38f1120f0e5d",
            "readOnly": true
          },
          "userId": {
            "$ref": "#/components/schemas/billing_id",
            "readOnly": true
          },
          "courseId": {
            "type": "string",
            "format": "uuid",
            "example": "0f4f19b1-1602-4c12-ac1a-9ad5baf1e4dc"
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "example": 499000,
            "readOnly": true
          },
          "status": {
            "$ref": "#/components/schemas/billing_TransactionStatus",
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-25T08:30:00Z",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-25T08:30:00Z",
            "readOnly": true
          }
        },
        "required": [
          "id",
          "userId",
          "courseId",
          "amount",
          "status",
          "updatedAt",
          "createdAt"
        ]
      },
      "billing_Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number",
            "minimum": 1,
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page",
            "minimum": 1,
            "maximum": 100,
            "example": 20
          },
          "total": {
            "type": "integer",
            "description": "Total number of items",
            "minimum": 0,
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages",
            "minimum": 0,
            "example": 5
          },
          "hasNext": {
            "type": "boolean",
            "description": "Whether there is a next page",
            "example": true
          },
          "hasPrev": {
            "type": "boolean",
            "description": "Whether there is a previous page",
            "example": false
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNext",
          "hasPrev"
        ]
      },
...
    },
    "responses": {
...
      "billing_BadRequestError": {
        "description": "Bad Request Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/billing_Error"
            },
            "example": {
              "code": "BAD_REQUEST",
              "message": "The request could not be understood by the server due to malformed syntax",
              "more_info": "https://api.egolia.com/docs/errors/BAD_REQUEST"
            }
          }
        }
      },
      "billing_UnauthorizedError": {
        "description": "Unauthorized Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "billing_ForbiddenError": {
        "description": "Forbidden Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/billing_Error"
            },
            "example": {
              "code": "FORBIDDEN",
              "message": "You do not have permission to access this resource",
              "more_info": "https://api.egolia.com/docs/errors/FORBIDDEN"
            }
          }
        }
      },
      "billing_InternalServerError": {
        "description": "Internal Server Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/billing_Error"
            },
            "example": {
              "code": "INTERNAL_SERVER_ERROR",
              "message": "An internal server error occurred. Please try again later",
              "more_info": "https://api.egolia.com/docs/errors/INTERNAL_SERVER_ERROR"
            }
          }
        }
      },
...
    }
  },
...
}