# pkg/api/billing/server.gen.go
// Package billing provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package billing

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get platform revenue analytics
	// (GET /billing/admin/analytics/revenue)
	GetPlatformRevenueAnalytics(c *gin.Context, params GetPlatformRevenueAnalyticsParams)
	// Checkout course
	// (POST /billing/courses/{courseId}/checkout)
	CheckoutCourse(c *gin.Context, courseId CourseIdPath)
	// Get transactions
	// (GET /billing/transactions)
	GetTransactions(c *gin.Context, params GetTransactionsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetPlatformRevenueAnalytics operation middleware
func (siw *ServerInterfaceWrapper) GetPlatformRevenueAnalytics(c *gin.Context) {

	var err error
	_ = err

	c.Set(string(Oauth2Scopes), []string{"openid", "entitlements"})

	c.Set(string(OIDCScopes), []string{"openid", "entitlements"})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetPlatformRevenueAnalyticsParams

	// ------------- Required query parameter "from" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "from", c.Request.URL.Query(), &params.From, runtime.BindQueryParameterOptions{Type: "string", Format: "date"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter from: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "to" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "to", c.Request.URL.Query(), &params.To, runtime.BindQueryParameterOptions{Type: "string", Format: "date"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter to: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetPlatformRevenueAnalytics(c, params)
}

// CheckoutCourse operation middleware
func (siw *ServerInterfaceWrapper) CheckoutCourse(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "courseId" -------------
	var courseId CourseIdPath

	err = runtime.BindStyledParameterWithOptions("simple", "courseId", c.Param("courseId"), &courseId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter courseId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(string(Oauth2Scopes), []string{"openid", "entitlements"})

	c.Set(string(OIDCScopes), []string{"openid", "entitlements"})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CheckoutCourse(c, courseId)
}

// GetTransactions operation middleware
func (siw *ServerInterfaceWrapper) GetTransactions(c *gin.Context) {

	var err error
	_ = err

	c.Set(string(Oauth2Scopes), []string{"openid", "entitlements"})

	c.Set(string(OIDCScopes), []string{"openid", "entitlements"})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTransactionsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page", c.Request.URL.Query(), &params.Page, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", c.Request.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "order" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "order", c.Request.URL.Query(), &params.Order, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter order: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "courseId" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "courseId", c.Request.URL.Query(), &params.CourseId, runtime.BindQueryParameterOptions{Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter courseId: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "learnerId" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "learnerId", c.Request.URL.Query(), &params.LearnerId, runtime.BindQueryParameterOptions{Type: "string", Format: "number"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter learnerId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetTransactions(c, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/billing/admin/analytics/revenue", wrapper.GetPlatformRevenueAnalytics)
	router.POST(options.BaseURL+"/billing/courses/:courseId/checkout", wrapper.CheckoutCourse)
	router.GET(options.BaseURL+"/billing/transactions", wrapper.GetTransactions)
}

type BadRequestErrorJSONResponse Error

type ForbiddenErrorJSONResponse Error

type InternalServerErrorJSONResponse Error

type UnauthorizedErrorJSONResponse map[string]interface{}

type GetPlatformRevenueAnalyticsRequestObject struct {
	Params GetPlatformRevenueAnalyticsParams
}

type GetPlatformRevenueAnalyticsResponseObject interface {
	VisitGetPlatformRevenueAnalyticsResponse(w http.ResponseWriter) error
}

type GetPlatformRevenueAnalytics200JSONResponse RevenueAnalytics

func (response GetPlatformRevenueAnalytics200JSONResponse) VisitGetPlatformRevenueAnalyticsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetPlatformRevenueAnalytics401JSONResponse struct{ UnauthorizedErrorJSONResponse }

func (response GetPlatformRevenueAnalytics401JSONResponse) VisitGetPlatformRevenueAnalyticsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetPlatformRevenueAnalytics403JSONResponse struct{ ForbiddenErrorJSONResponse }

func (response GetPlatformRevenueAnalytics403JSONResponse) VisitGetPlatformRevenueAnalyticsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type GetPlatformRevenueAnalytics500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response GetPlatformRevenueAnalytics500JSONResponse) VisitGetPlatformRevenueAnalyticsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type CheckoutCourseRequestObject struct {
	CourseId CourseIdPath `json:"courseId"`
}

type CheckoutCourseResponseObject interface {
	VisitCheckoutCourseResponse(w http.ResponseWriter) error
}

type CheckoutCourse201JSONResponse struct {
	PaymentUrl    string             `json:"paymentUrl"`
	TransactionId openapi_types.UUID `json:"transactionId"`
}

func (response CheckoutCourse201JSONResponse) VisitCheckoutCourseResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type CheckoutCourse400JSONResponse struct{ BadRequestErrorJSONResponse }

func (response CheckoutCourse400JSONResponse) VisitCheckoutCourseResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type CheckoutCourse401JSONResponse struct{ UnauthorizedErrorJSONResponse }

func (response CheckoutCourse401JSONResponse) VisitCheckoutCourseResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type CheckoutCourse403JSONResponse struct{ ForbiddenErrorJSONResponse }

func (response CheckoutCourse403JSONResponse) VisitCheckoutCourseResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type CheckoutCourse500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response CheckoutCourse500JSONResponse) VisitCheckoutCourseResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetTransactionsRequestObject struct {
	Params GetTransactionsParams
}

type GetTransactionsResponseObject interface {
	VisitGetTransactionsResponse(w http.ResponseWriter) error
}

type GetTransactions200JSONResponse struct {
	Data       []Transaction `json:"data"`
	Pagination Pagination    `json:"pagination"`
}

func (response GetTransactions200JSONResponse) VisitGetTransactionsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetTransactions401JSONResponse struct{ UnauthorizedErrorJSONResponse }

func (response GetTransactions401JSONResponse) VisitGetTransactionsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetTransactions500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response GetTransactions500JSONResponse) VisitGetTransactionsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Get platform revenue analytics
	// (GET /billing/admin/analytics/revenue)
	GetPlatformRevenueAnalytics(ctx context.Context, request GetPlatformRevenueAnalyticsRequestObject) (GetPlatformRevenueAnalyticsResponseObject, error)
	// Checkout course
	// (POST /billing/courses/{courseId}/checkout)
	CheckoutCourse(ctx context.Context, request CheckoutCourseRequestObject) (CheckoutCourseResponseObject, error)
	// Get transactions
	// (GET /billing/transactions)
	GetTransactions(ctx context.Context, request GetTransactionsRequestObject) (GetTransactionsResponseObject, error)
}

type StrictHandlerFunc func(ctx *gin.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictGinServerOptions struct {
	// RequestErrorHandlerFunc is called when a request cannot be parsed or
	// decoded. It is invoked for JSON bind failures, form parse/bind errors,
	// multipart reader errors, media type parse errors, missing multipart
	// boundaries, and request body read errors. The default returns 400.
	RequestErrorHandlerFunc func(ctx *gin.Context, err error)
	// HandlerErrorFunc is called when the application handler (or any
	// middleware wrapping it) returns a non-nil error. The default returns 500.
	HandlerErrorFunc func(ctx *gin.Context, err error)
	// ResponseErrorHandlerFunc is called when the response object fails to
	// serialize (Visit*Response returns an error) or when the handler returns
	// an unexpected response type. The default returns 500.
	ResponseErrorHandlerFunc func(ctx *gin.Context, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictGinServerOptions{
		RequestErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		},
		HandlerErrorFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
		ResponseErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictGinServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		}
	}
	if options.HandlerErrorFunc == nil {
		options.HandlerErrorFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictGinServerOptions
}

// GetPlatformRevenueAnalytics operation middleware
func (sh *strictHandler) GetPlatformRevenueAnalytics(ctx *gin.Context, params GetPlatformRevenueAnalyticsParams) {
	var request GetPlatformRevenueAnalyticsRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetPlatformRevenueAnalytics(ctx, request.(GetPlatformRevenueAnalyticsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetPlatformRevenueAnalytics")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetPlatformRevenueAnalyticsResponseObject); ok {
		if err := validResponse.VisitGetPlatformRevenueAnalyticsResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CheckoutCourse operation middleware
func (sh *strictHandler) CheckoutCourse(ctx *gin.Context, courseId CourseIdPath) {
	var request CheckoutCourseRequestObject

	request.CourseId = courseId

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CheckoutCourse(ctx, request.(CheckoutCourseRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CheckoutCourse")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(CheckoutCourseResponseObject); ok {
		if err := validResponse.VisitCheckoutCourseResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetTransactions operation middleware
func (sh *strictHandler) GetTransactions(ctx *gin.Context, params GetTransactionsParams) {
	var request GetTransactionsRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetTransactions(ctx, request.(GetTransactionsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTransactions")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetTransactionsResponseObject); ok {
		if err := validResponse.VisitGetTransactionsResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}